backend/requirements.txt
fastapi
uvicorn
crewai
langgraph
langfuse
python-dotenv
PyGithub
langchain-openai
pydantic
langchain
langchain-community
